refactor: util 함수 파일 정리
